+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## v0.2.5 - 2017-05-03
+### Added:
+ - Negotiations: invited, view, comments for employer, send a message
+ - Suggests
+ - Request locale, host
+
 ## v0.2.4 -  2017-05-01
-### Added
+### Added:
  - Dictionaries endpoint
  - Comments: create, edit, delete